repay poc (#524)
* repay poc

* Integrate wasm source more closely with margin package (#527)

* Fixes quote_to_base

* Adds console_log!, sets panic hook

Weirdly, init_module does not appear to get called on load. Leaving
that to be figured out later.

* Integrates wasm module more closely with margin package

* Remove testing snippet

* client: split out instruction builders to another crate (#526)

* upgrade pyth library (#481)

* add extra exports from program crates (#530)

- fixed-term now exports the `origination_fee` module
- margin-pool now exports the `PoolAction` type, which is used in
  conversion functions on the pool

* upgrade simulator (#532)

* sdk: minor refactoring (#533)

- require airspace for margin ix builder
- simplify API for building swap instructions
- stop returning nonces when deriving addreses, since they are never
  really required

* allow jet-lab-wasm-src to depend on jet-fixed-term (#529)

* Unused fature on wasm-bindgen is causing cyclical deps (#536)

* Emit margin config events (#535)

* remove unused var

+++ libraries/rust/instructions/src/fixed_term.rs
@@ -766,7 +766,7 @@ impl FixedTermIxBuilder {
         self.modify_market([false as u8].into(), 8 + 32 * 14 + 2)
     }
 
-    pub fn modify_market(&self, data: Vec<u8>, offset: usize) -> Instruction {
+    pub fn modify_market(&self, data: Vec<u8>, offset: u32) -> Instruction {
         let data = jet_fixed_term::instruction::ModifyMarket { data, offset }.data();
         let accounts = jet_fixed_term::accounts::ModifyMarket {
             market: self.market,
